requirements.txt
django==5.2
python-dotenv==1.1.0
djangorestframework==3.16.0
drf-spectacular==0.28.0